Reworked xml loading and saving so all characters are stored in one file. Also added the code to properly save and delete multiple characters.

+++ DnD/Assets/Scripts/Data Handling Scripts/Data_Loader.cs
@@ -9,10 +9,64 @@ using System.Text;
 using System.Security.Cryptography;
 
 public class Data_Loader : ScriptableObject {
-	public void LoadCharacterData(string filename)
+    public List<string> LoadCharacterIDs()
+    {
+        string line = "";
+        string input_file = "./Saved Data/characters.xml";
+        string output_file = "temp.xml";
+        List<string> keyList = new List<string>();
+        List<string> elemList = new List<string>();
+        UnicodeEncoding encoding = new UnicodeEncoding();
+        byte[] key = null;
+        RijndaelManaged RMCrypto = new RijndaelManaged();
+        Image_Converter ImageConverter = ScriptableObject.CreateInstance<Image_Converter>();
+        int i = 0;
+
+        elemList.Clear();
+
+        //Get encryption / decryption key from url
+        XML_Loader XML = ScriptableObject.CreateInstance<XML_Loader>();
+        key = encoding.GetBytes(Data_Handler_Key.keyvalue);
+        if (Directory.Exists(Path.Combine(Directory.GetCurrentDirectory(), "Saved Data")) && (File.Exists((Path.Combine(Directory.GetCurrentDirectory(), "Saved Data/characters.xml")))))
+        {
+            //Open / read the selected (encrypted) character file and decrypt it, then write the decrypted information to a temporary xml file
+            FileStream decrypted_file = new FileStream(input_file, FileMode.Open);
+            FileStream temp_file = new FileStream(output_file, FileMode.Create);
+            CryptoStream cryptography_stream = new CryptoStream(decrypted_file, RMCrypto.CreateDecryptor(key, key), CryptoStreamMode.Read);
+            using (MemoryStream msDecrypt = new MemoryStream())
+            {
+                using (StreamReader srDecrypt = new StreamReader(cryptography_stream))
+                {
+                    using (StreamWriter swTemp = new StreamWriter(temp_file))
+                    {
+                        while ((line = srDecrypt.ReadLine()) != null)
+                        {
+                            swTemp.WriteLine(line);
+                        }
+                    }
+                }
+            }
+            cryptography_stream.Close();
+            decrypted_file.Close();
+
+            //Call functions to load data from temporary xml file into specified game objects
+            elemList = XML.LoadInnerXmlFromFile(output_file, "id");
+            foreach (var item in elemList)
+            {
+                if (i < Convert.ToInt32(item))
+                    i = Convert.ToInt32(item);
+            }
+
+            //Delete the temporary xml file
+            File.Delete(output_file);
+        }
+        Character_Info.maxid = i + 1;
+        return elemList;
+    }
+	public void LoadCharacterData(int id)
 	{
 		string line = "";
-		string input_file = filename;
+		string input_file = "./Saved Data/characters.xml";
 		string output_file = "temp.xml";
 		List<string> keyList = new List<string>();
 		List<string> elemList = new List<string> ();
@@ -20,6 +74,7 @@ public class Data_Loader : ScriptableObject {
 		byte[] key = null;
 		RijndaelManaged RMCrypto = new RijndaelManaged ();
         Image_Converter ImageConverter = ScriptableObject.CreateInstance<Image_Converter>();
+        string tagName = "";
 		
 		//Get encryption / decryption key from url
 		XML_Loader XML = ScriptableObject.CreateInstance<XML_Loader> ();
@@ -42,117 +97,106 @@ public class Data_Loader : ScriptableObject {
 		decrypted_file.Close ();
 
 		//Call functions to load data from temporary xml file into specified game objects
-        elemList = XML.LoadInnerXmlFromFile(output_file, "avatar");
-        foreach (var item in elemList)
-        {
-            Character_Info.characterAvatar = ImageConverter.ConvertStringToImage(item);
-        }
+        tagName = "avatar" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+        Character_Info.characterAvatar = ImageConverter.ConvertStringToImage(elemList[0]);
         elemList.Clear();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "charactername");
-		foreach(var item in elemList){
-			Character_Info.characterName = item;
-		}
+        tagName = "charactername" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+		Character_Info.characterName = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "characterclass");
-		foreach(var item in elemList){
-			Character_Info.characterClass = item;
-		}
+        tagName = "characterclass" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+		Character_Info.characterClass = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "characterrace");
-		foreach(var item in elemList){
-			Character_Info.characterRace = item;
-		}
+        tagName = "characterrace" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+        Character_Info.characterRace = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "charactersubrace");
-		foreach(var item in elemList){
-			Character_Info.characterSubrace = item;
-		}
+        tagName = "charactersubrace" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+	    Character_Info.characterSubrace = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "alignment");
-		foreach(var item in elemList){
-			Character_Info.characterAlignment = item;
-		}
+        tagName = "alignment" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+	    Character_Info.characterAlignment = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "age");
-		foreach(var item in elemList){
-			Character_Info.characterAge = item;
-		}
+		tagName = "age" + id;
+		elemList = XML.LoadInnerXmlFromFile (output_file, tagName);
+	    Character_Info.characterAge = elemList[0];
+        elemList.Clear();
+
+        tagName = "gender" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+	    Character_Info.characterGender = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "gender");
-		foreach(var item in elemList){
-			Character_Info.characterGender = item;
-		}
+        tagName = "characterlevel" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+        Character_Info.characterLevel = elemList[0];
+        elemList.Clear();
+
+        tagName = "health" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+        Character_Info.characterHealth = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "characterlevel");
-		foreach(var item in elemList){
-			Character_Info.characterLevel = item;
-		}
+        tagName = "height" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+		Character_Info.characterHeight = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "health");
-		foreach(var item in elemList){
-			Character_Info.characterHealth = item;
-		}
+        tagName = "weight" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+		Character_Info.characterWeight = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "height");
-		foreach(var item in elemList){
-			Character_Info.characterHeight = item;
-		}
+        tagName = "carryweight" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+        Character_Info.characterCarryWeight = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "weight");
-		foreach(var item in elemList){
-			Character_Info.characterWeight = item;
-		}
+        tagName = "movespeed" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+		Character_Info.characterMoveSpeed = elemList[0];
 		elemList.Clear ();
 
-		elemList = XML.LoadInnerXmlFromFile (output_file, "carryweight");
-		foreach(var item in elemList){
-			Character_Info.characterCarryWeight = item;
-		}
-		elemList.Clear ();
-
-		elemList = XML.LoadInnerXmlFromFile (output_file, "movespeed");
-		foreach(var item in elemList){
-			Character_Info.characterMoveSpeed = item;
-		}
-		elemList.Clear ();
-
-		elemList = XML.LoadInnerXmlFromFile (output_file, "languages");
-		foreach(var item in elemList){
-			Character_Info.characterLanguages = item;
-		}
+        tagName = "languages" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
+		Character_Info.characterLanguages = elemList[0];
 		elemList.Clear ();
 
 		Note_List_Info.noteTitles.Clear ();
-		elemList = XML.LoadInnerXmlFromFile (output_file, "title");
+        tagName = "title" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
 		foreach(var item in elemList){
 			Note_List_Info.noteTitles.Add (item);
 		}
 		elemList.Clear ();
 
 		Note_List_Info.noteDates.Clear ();
-		elemList = XML.LoadInnerXmlFromFile (output_file, "date");
+        tagName = "date" + id;
+        elemList = XML.LoadInnerXmlFromFile(output_file, tagName);
 		foreach(var item in elemList){
 			Note_List_Info.noteDates.Add (item);
 		}
 		elemList.Clear ();
 
 		Note_List_Info.noteSubjects.Clear ();
-		elemList = XML.LoadInnerXmlFromFile (output_file, "subject");
+        tagName = "subject" + id;
+		elemList = XML.LoadInnerXmlFromFile (output_file, tagName);
 		foreach(var item in elemList){
 			Note_List_Info.noteSubjects.Add(item);
 		}
 		elemList.Clear ();
+
+        Character_Info.id = id;
 
 		//Delete the temporary xml file
 		File.Delete (output_file);
@@ -162,7 +206,7 @@ public class Data_Loader : ScriptableObject {
     public void LoadSettingsData()
     {
         string line = "";
-        string input_file = "./Settings/settings.xml";
+        string input_file = "./Saved Data/settings.xml";
         string output_file = "temp.xml";
         List<string> keyList = new List<string>();
         List<string> elemList = new List<string>();
@@ -175,42 +219,45 @@ public class Data_Loader : ScriptableObject {
         key = encoding.GetBytes(Data_Handler_Key.keyvalue);
 
         //Open / read the selected (encrypted) character file and decrypt it, then write the decrypted information to a temporary xml file
-        FileStream decrypted_file = new FileStream(input_file, FileMode.Open);
-        FileStream temp_file = new FileStream(output_file, FileMode.Create);
-        CryptoStream cryptography_stream = new CryptoStream(decrypted_file, RMCrypto.CreateDecryptor(key, key), CryptoStreamMode.Read);
-        using (MemoryStream msDecrypt = new MemoryStream())
+        if (Directory.Exists(Path.Combine(Directory.GetCurrentDirectory(), "Saved Data")) && (File.Exists((Path.Combine(Directory.GetCurrentDirectory(), "Saved Data/settings.xml")))))
         {
-            using (StreamReader srDecrypt = new StreamReader(cryptography_stream))
+            FileStream decrypted_file = new FileStream(input_file, FileMode.Open);
+            FileStream temp_file = new FileStream(output_file, FileMode.Create);
+            CryptoStream cryptography_stream = new CryptoStream(decrypted_file, RMCrypto.CreateDecryptor(key, key), CryptoStreamMode.Read);
+            using (MemoryStream msDecrypt = new MemoryStream())
             {
-                using (StreamWriter swTemp = new StreamWriter(temp_file))
+                using (StreamReader srDecrypt = new StreamReader(cryptography_stream))
                 {
-                    while ((line = srDecrypt.ReadLine()) != null)
+                    using (StreamWriter swTemp = new StreamWriter(temp_file))
                     {
-                        swTemp.WriteLine(line);
+                        while ((line = srDecrypt.ReadLine()) != null)
+                        {
+                            swTemp.WriteLine(line);
+                        }
                     }
                 }
             }
-        }
-        cryptography_stream.Close();
-        decrypted_file.Close();
+            cryptography_stream.Close();
+            decrypted_file.Close();
 
-        //Call functions to load data from temporary xml file into specified game objects
-        elemList = XML.LoadInnerXmlFromFile(output_file, "mode");
-        foreach (var item in elemList)
-        {
-            switch(item)
+            //Call functions to load data from temporary xml file into specified game objects
+            elemList = XML.LoadInnerXmlFromFile(output_file, "mode");
+            foreach (var item in elemList)
             {
-                case "true":
-                    Settings_Screen.is_online = true;
-                    break;
-                case "false":
-                    Settings_Screen.is_online = false;
-                    break;
+                switch (item)
+                {
+                    case "true":
+                        Settings_Screen.is_online = true;
+                        break;
+                    case "false":
+                        Settings_Screen.is_online = false;
+                        break;
+                }
             }
-        }
-        elemList.Clear();
+            elemList.Clear();
 
-        //Delete the temporary xml file
-        File.Delete(output_file);
+            //Delete the temporary xml file
+            File.Delete(output_file);
+        }
     }
 }
